Switch LVQ to cascaded scalar quantization (#312)
We've decided to switch the implementation of two-level LVQ from its original version (which basically considered the residual as additional bits appended to the primary encoding) to true cascaded scalar quantization.

Let
* $P$ be the number of bits used for the primary encoding.
* $R$ be the number of bits used for the residual encoding.
* $\Delta$ be the scaling parameter for the primary encoding (difference between maximum and minimum values).

Then the expected error for two-level LVQ goes from
```math
\frac{\Delta}{2^P - 1} \cdot \frac{1}{2 \cdot 2^R}
```
with the caveat that some values with the maximum encoding for the residual may have twice this error to
```math
\frac{\Delta}{2^P - 1} \cdot \frac{1}{2 \cdot (2^R - 1)}
```
without that caveat.

Benchmarking suggests that there is minimal performance gain or loss due to this change. We're basically making it to be consistent with the VLDB paper.

GitOrigin-RevId: 6231847932dc3ef6a59adece47bd03dabbdd2b95