Fix a performance pathology in dynamic LVQ4(xN) (#332)
Hitting the predicated epilogue for dynamically dimensioned LVQ 4(xN) kernel computations turns out to be incredibly slow due to variable length calls to `std::memcpy`. Fix that by using a predicated SIMD load (relying on fault suppression to avoid out-of-bounds reads).

In the static case, try to not use this predicated load as the compiler is able to constant-propagate the number of bytes to read and generally the inlined call to `std::memcpy` ends up being slightly faster.

GitOrigin-RevId: 79a2324fa46a51e22a68f0ad1453ea4bee06a0c5