Solve batch iterator reranking issue (#95)
This PR addresses the reranking issue in the batch iterator. When the
dataset requires reranking after a search, it is necessary to restore
the distances using the primary dataset before proceeding to search for
the next batch. Failing to do so may result in the search buffer being
mixed with both reranking results and primary dataset results.